% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gantt_DiagrammeR.R
\name{gantt_DiagrammeR_chart}
\alias{gantt_DiagrammeR_chart}
\title{用于生成甘特图}
\usage{
gantt_DiagrammeR_chart(title = "这是示例甘特图",
  dataFileName = "~/pkgs/tsdv/inst/extdata/gantt_dynamic_data.xlsx")
}
\arguments{
\item{title}{标题}

\item{dataFileName}{文件}
}
\value{
返回值
}
\description{
用于生成甘特图
}
\examples{
gantt_DiagrammeR_chart();
}
