% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gantt_DiagrammeR_data.R
\name{gantt_DiagrammeR_data}
\alias{gantt_DiagrammeR_data}
\title{用于diagrammeR的甘特图的数据准备}
\usage{

  gantt_DiagrammeR_data(file = "~/pkgs/tsdv/inst/extdata/gantt_dynamic_data.xlsx")
}
\arguments{
\item{file}{文件名}
}
\value{
返回值
}
\description{
用于diagrammeR的甘特图的数据准备
}
\examples{
gantt_DiagrammeR_data();
}
