% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kBar.recharts3.dim.stack.R
\name{kBar.recharts3.dim.stack}
\alias{kBar.recharts3.dim.stack}
\title{插入柱状图-维度数据堆叠}
\usage{
kBar.recharts3.dim.stack(data = kBar_dim_data, x = "seller",
  y = "saleNum", z = "weekDay")
}
\arguments{
\item{data}{数据源}

\item{x}{第一个维度}

\item{y}{第二个维度}

\item{z}{第三个维度，也指series.}
}
\value{
返回值
}
\description{
插入柱状图-维度数据堆叠
}
\examples{
kBar.recharts3.dim.stack();
}
