% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kBar.recharts3.obj.horiz.R
\name{kBar.recharts3.obj.horiz}
\alias{kBar.recharts3.obj.horiz}
\title{插入条形图-默认数据列}
\usage{
kBar.recharts3.obj.horiz(data = kBar_obj_data)
}
\arguments{
\item{data}{数据源}
}
\value{
返回值
}
\description{
插入条形图-默认数据列
}
\examples{
kBar.recharts3.obj.horiz();
}
