% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kBar.recharts3.obj.R
\name{kBar.recharts3.obj}
\alias{kBar.recharts3.obj}
\title{插入柱状图-默认数据}
\usage{
kBar.recharts3.obj(data = kBar_obj_data, horiz = FALSE)
}
\arguments{
\item{data}{数据源}

\item{horiz}{默认为柱状图，设置为FALSE,否则为条形图}
}
\value{
返回值
}
\description{
插入柱状图-默认数据
}
\examples{
kBar.recharts3.obj();
}
