% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kBar.recharts3.rnorm.R
\name{kBar.recharts3.rnorm}
\alias{kBar.recharts3.rnorm}
\title{插入柱状图-正态分布}
\usage{
kBar.recharts3.rnorm(data = kBar_rnorm_data)
}
\arguments{
\item{data}{数据源}
}
\value{
返回值
}
\description{
插入柱状图-正态分布
}
\examples{
kBar.recharts3.rnorm();
}
