% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kLine.recharts3.dim.R
\name{kLine.recharts3.dim}
\alias{kLine.recharts3.dim}
\title{插入线图-维度数据}
\usage{
kLine.recharts3.dim(data = df2, x = "weekDay", y = "saleNum",
  z = "seller")
}
\arguments{
\item{data}{数据源}

\item{x}{第一个维度}

\item{y}{第二个维度}

\item{z}{第三个维度，也可以是series}
}
\value{
返回值
}
\description{
插入线图-维度数据
}
\examples{
kLine.recharts3.dim();
}
