% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kLine.recharts3.obj.R
\name{kLine.recharts3.obj}
\alias{kLine.recharts3.obj}
\title{插入线图-基本}
\usage{
kLine.recharts3.obj(data = WorldPhones, theme = 1)
}
\arguments{
\item{data}{数据集}

\item{theme}{主题色系}
}
\value{
返回值
}
\description{
插入线图-基本
}
\examples{
kLine.recharts3.obj();
}
