% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kLine.recharts3.rnorm.R
\name{kLine.recharts3.rnorm}
\alias{kLine.recharts3.rnorm}
\title{插入线图-正态分布}
\usage{
kLine.recharts3.rnorm(data = kline_rnorm_data)
}
\arguments{
\item{data}{数据源}
}
\value{
返回值
}
\description{
插入线图-正态分布
}
\examples{
kLine.rechats3.rnorm();
}
