% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kPie.recharts3.dim.R
\name{kPie.recharts3.dim}
\alias{kPie.recharts3.dim}
\title{插入饼图-维度数据}
\usage{
kPie.recharts3.dim(data = kPie_dim_data, x = "stdName", y = "val1")
}
\arguments{
\item{data}{数据源}

\item{x}{name var.}

\item{y}{value var.}
}
\value{
返回值
}
\description{
插入饼图-维度数据
}
\examples{
kPie.recharts3.dim();
}
