% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kPie.recharts3.obj.R
\name{kPie.recharts3.obj}
\alias{kPie.recharts3.obj}
\title{插入饼图-默认数据列}
\usage{
kPie.recharts3.obj(data = kPie_obj_data, title = "test")
}
\arguments{
\item{data}{数据源}

\item{title}{标题}
}
\value{
返回值
}
\description{
插入饼图-默认数据列
}
\examples{
kPie.recharts3.obj();
}
