% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kPoint.recharts3.dim.R
\name{kPoint.recharts3.dim}
\alias{kPoint.recharts3.dim}
\title{插入散步图，按维度指定数据列}
\usage{
kPoint.recharts3.dim(data = iris[, 3:5], x = "Petal.Length",
  y = "Petal.Width", z = "Species", theme = 1)
}
\arguments{
\item{data}{数据源}

\item{x}{X轴数据}

\item{y}{Y辆数据}

\item{z}{Z轴数据，也称为分层数据}

\item{theme}{主题}
}
\value{
返回值
}
\description{
插入散步图，按维度指定数据列
}
\examples{
kPoint.recharts3.dim();
}
