% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kPoint.recharts3.obj.R
\name{kPoint.recharts3.obj}
\alias{kPoint.recharts3.obj}
\title{插入recharts3的默认表数据}
\usage{
kPoint.recharts3.obj(data = iris[, 3:5], theme = 2)
}
\arguments{
\item{data}{输入数据}

\item{theme}{输入主题代码}
}
\value{
返回值
}
\description{
插入recharts3的默认表数据
}
\examples{
kPoint.recharts3.obj();
}
